extract ai conversation related hook
